Replace previous printSchema with updated version
Copy over the now validated `newPrintFederatedSchema.ts` contents
into the former `printFederatedSchema.ts` as an "under-the-hood"
swap.
